+++ src/containerapp/azext_containerapp/_models.py
@@ -526,6 +526,11 @@ ExtendedLocation = {
     "type": None
 }
 
+ManagedServiceIdentity = {
+    "type": None,  # 'None', 'SystemAssigned', 'UserAssigned', 'SystemAssigned,UserAssigned'
+    "userAssignedIdentities": None  # {string: UserAssignedIdentity}
+}
+
 JavaComponent = {
     "properties": {
         "componentType": None,